// src/API.ts
export enum TxType {
  SEND = "SEND",
  REQUEST = "REQUEST",
  SEND_REQUEST = "SEND_REQUEST",
  TX_REJECTED = "TX_REJECTED",
  DOWN_MONEY_CASH = "DOWN_MONEY_CASH",
  UP_MONEY_CASH = "UP_MONEY_CASH",
}


export enum TxStatus {
  SEND = "SEND",
  REQUEST = "REQUEST",
  STANDBY = "STANDBY",
  SHARED = "SHARED",
  CONFIRM = "CONFIRM",
  DOWNLOAD = "DOWNLOAD",
  REJECT = "REJECT",
  UP_CASH = "UP_CASH",
  DOWN_CASH = "DOWN_CASH",
}